+++ lib/rspec/core/rake_task.rb
@@ -45,6 +45,9 @@ module RSpec
       # A message to print to stderr when there are failures.
       attr_accessor :failure_message
 
+      # Run RSpec with a clean (empty) environment.
+      attr_accessor :with_clean_environment
+
       # Use verbose output. If this is set to true, the task will print the
       # executed spec command to stdout. Defaults to `true`.
       attr_accessor :verbose
@@ -76,7 +79,12 @@ module RSpec
         command = spec_command
         puts command if verbose
 
-        return if system(command)
+        if with_clean_environment
+          return if system({}, command, :unsetenv_others => true)
+        else
+          return if system(command)
+        end
+
         puts failure_message if failure_message
 
         return unless fail_on_error